Added two screens for adding incomes and expenses and made them work with. Currently HomeScreen does not update state even though DB is updated

+++ app/src/main/java/com/example/financemanagementappv2/data/entities/Incomes.kt
@@ -19,5 +19,5 @@ data class Incomes (
     val amount: Double,
     val description: String,
     val date: Long,
-    val categoryId: Int,
+    val categoryId: Long,
 )
